+++ src/dreamstate/core/consumption/useManager.ts
@@ -29,9 +29,7 @@ export function useManager<T extends TAnyObject, D extends IContextManagerConstr
    * Fallback to pub-sub + checking approach only if dependency selector was provided.
    * If component should update on every change there is no point of doing anything additional to default context.
    */
-  if (dependenciesSelector) {
-    return useContextWithMemo<T, D>(ManagerClass, dependenciesSelector);
-  } else {
-    return useContext(ManagerClass.REACT_CONTEXT);
-  }
+  return dependenciesSelector
+    ? useContextWithMemo<T, D>(ManagerClass, dependenciesSelector)
+    : useContext(ManagerClass.REACT_CONTEXT);
 }